Decomp SetLeaderAction in overlay_29
Split asm/overlay_29_022F05B4.s at 0x022F2B3C; add src/overlay_29_022F0EDC.c
and its header, and both objects to main.lsf.

Type corrections the match required:
- dungeon.field_0x1d8 / field_0x1dc: paired u16 fields -> struct position
- dungeon.field_0x614: u32 -> s32, tested >= 0
- display_data.leader_target_direction_mirror: enum direction_id -> u8; the
  enum type folds the stored 0xFF to -1 and drops retail's mov r0, #0xff
- CanMonsterMoveInDirection's direction parameter: u16 -> s32; a u16 parameter
  narrows the argument at the new call site, where retail passes it unchanged

Authored by Claude (Opus 5) under human direction. Confirmed by a matching
build: build/pmdsky.us/pmdsky.us.nds: OK.

Not verified: the JP target, which is expected not to match -- the replaced asm
carries JAPAN-specific variants beyond the message-id offset. Two volatile
stand-ins remain in the new source, and this translation unit declares
DUNGEON_PTR as a scalar where other files declare it as an array; the two are
not byte-interchangeable here.

+++ include/dungeon_capabilities_3.h
@@ -10,7 +10,7 @@
 bool8 MonsterCannotAttack(struct entity *entity, bool8 skip_sleep);
 // Checks if the given monster can move in the specified direction
 // Returns false if any monster is standing on the target tile
-bool8 CanMonsterMoveInDirection(struct entity *monster, u16 direction);
+bool8 CanMonsterMoveInDirection(struct entity *monster, s32 direction);
 #ifndef JAPAN
 // Returns the mobility type of a monster, after accounting for things that could affect it.
 // List of checks: Mobile status, Mobile Scarf, All-Terrain Hiker and Absolute Mover.
